Move uint_from_xof Clippy fixes
