+++ synedrion/src/tools/hashing.rs
@@ -1,4 +1,4 @@
-use digest::{Digest, ExtendableOutput, Update};
+use digest::{Digest, ExtendableOutput, Update, XofReader};
 use serde::{Deserialize, Serialize};
 use sha2::Sha256;
 use sha3::{Shake256, Shake256Reader};
@@ -7,6 +7,7 @@ use hashing_serializer::HashingSerializer;
 
 use crate::curve::Scalar;
 use crate::tools::serde_bytes;
+use crypto_bigint::{Encoding, Integer, NonZero};
 
 /// A digest object that takes byte slices or decomposable ([`Hashable`]) objects.
 pub trait Chain: Sized {
@@ -145,5 +146,36 @@ impl<T: Serialize> Hashable for T {
             .expect("The type is serializable");
 
         digest
+    }
+}
+
+/// Build a `T` integer from an extendable Reader function
+pub(crate) fn uint_from_xof<T>(reader: &mut impl XofReader, modulus: &NonZero<T>) -> T
+where
+    T: Integer + Encoding,
+{
+    let backend_modulus = modulus.as_ref();
+
+    let n_bits = backend_modulus.bits_vartime();
+    let n_bytes = (n_bits + 7) / 8; // ceiling division by 8
+
+    // If the number of bits is not a multiple of 8,
+    // use a mask to zeroize the high bits in the gererated random bytestring,
+    // so that we don't have to reject too much.
+    let mask = if n_bits & 7 != 0 {
+        (1 << (n_bits & 7)) - 1
+    } else {
+        u8::MAX
+    };
+
+    let mut bytes = T::zero().to_le_bytes();
+    loop {
+        reader.read(&mut (bytes.as_mut()[0..n_bytes as usize]));
+        bytes.as_mut()[n_bytes as usize - 1] &= mask;
+        let n = T::from_le_bytes(bytes);
+
+        if n.ct_lt(backend_modulus).into() {
+            return n;
+        }
     }
 }
