+++ synedrion/src/uint.rs
@@ -18,38 +18,6 @@ use crypto_bigint::{
     subtle::{ConditionallySelectable, CtOption},
     Square,
 };
-use digest::XofReader;
-
-/// Build a `T` integer from an extendable Reader function
-pub(crate) fn uint_from_xof<T>(reader: &mut impl XofReader, modulus: &NonZero<T>) -> T
-where
-    T: Integer + Encoding,
-{
-    let backend_modulus = modulus.as_ref();
-
-    let n_bits = backend_modulus.bits_vartime();
-    let n_bytes = (n_bits + 7) / 8; // ceiling division by 8
-
-    // If the number of bits is not a multiple of 8,
-    // use a mask to zeroize the high bits in the gererated random bytestring,
-    // so that we don't have to reject too much.
-    let mask = if n_bits & 7 != 0 {
-        (1 << (n_bits & 7)) - 1
-    } else {
-        u8::MAX
-    };
-
-    let mut bytes = T::zero().to_le_bytes();
-    loop {
-        reader.read(&mut (bytes.as_mut()[0..n_bytes as usize]));
-        bytes.as_mut()[n_bytes as usize - 1] &= mask;
-        let n = T::from_le_bytes(bytes);
-
-        if n.ct_lt(backend_modulus).into() {
-            return n;
-        }
-    }
-}
 
 pub(crate) fn pow_signed<T>(
     uint: <T as Integer>::Monty,
